Create index.php
In the index.php file, you will find the necessary to code to set up the connection between your Arduino Uno board and your new website. The code here will create an on and off button on your website, and will turn the LED light on your Arduino on and off when you click each button respectively. 
The code here can be altered to your custom needs but this will set up the initial connection on the PHP side.