Create Arduino code - LED light
The code you will find here is the second piece to setting up the connection with your Arduino. This code interacts with the PHP that is written in the index.php file and tells the light which pin to turn on and off when the button on your website is clicked.